pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "nltk>=3.9.1",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "streamlit>=1.42.0",
    "twilio>=9.4.4",
]
